pyproject.toml
[tool.poetry]
name = "reverse-engineering-agent"
version = "0.1.0"
description = "An agent for reverse-engineering data models from SQL code."
authors = ["Your Name <you@example.com>"] # Replace with your name and email
readme = "README.md"
package-mode = false


[tool.poetry.dependencies]
python = ">=3.9, !=3.9.7, <3.12" # Exclude Python 3.9.7 due to streamlit incompatibility
pandas = "^2.0.0" # Use an appropriate version
google-cloud-storage = "^2.0.0" # Use an appropriate version
pydantic-settings = "2.5.2"
google-cloud-aiplatform = "*"
google-cloud-pipeline-components = "2.3.0"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.30.1"}
streamlit = "^1.36.0"
requests = "^2.32.3"
pyarrow = "^16.1.0"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0.0" # Example dev dependency

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
